+++ private/write_nii.m
@@ -5,7 +5,7 @@ case 'float32'
     fa = file_array(f,size(img),typ,0);
 case 'uint8'
     mx = max(img(isfinite(img(:))));
-    fa = file_array(f,size(img),typ,0,mx/255,0);
+    fa = file_array(f,size(img),typ,0,double(mx)/255,0);
 case 'int16'
     mx = max(img(isfinite(img(:))));
     mn = min(img(isfinite(img(:))));
